Part2 completed

+++ src/main/java/se/lexicon/ishwarya/booklender_springboot/repository/LoanRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Repository
 public interface LoanRepository extends CrudRepository<Loan,Long> {
     List<Loan> findByUserId(Long userId);
-    List<Loan> findByBookId(Long bookId);
+    List<Loan> findByBookId(int bookId);
     @Query("SELECT l FROM Loan l WHERE l.concluded = :concluded")
     List<Loan> findByConcludedStatus(@Param("concluded") boolean concluded);
